Ref: Changed search filters, enchance plant profile and slider
- Refactored code to allow usage of plant in plants.jsx in test_folder
  for plant_profile.jsx.
- Updated plant search filters to focus on different criteria for improved search functionality.
- Modified plant slider to display image, plant name, citation, and price in USD.
- Added health benefits and pricing information on the plant profile page.
