+++ src/App.jsx
@@ -8,7 +8,6 @@ import PlantsSlider from './components/plants_slider'
 import FilterModal from './components/nav_bar/filter/filter_modal'
 import {plants} from "./components/test_data/plants"
 import PlantProfileCard from './components/plant_profile'
-import PlantProfileCardAnimated from './components/plant_profile_motion_version'
 
 
 function App() {
@@ -21,15 +20,9 @@ function App() {
     Cost: 0
   });
 
-  // Two Versions
-  /*
-  return (
-    <> <PlantProfileCard/></>
-  );
-  return (
-    <> <PlantProfileCardAnimated/></>
-  );
-  */
+  // return (
+  //   <> <PlantProfileCard/></>
+  // );
 
   return (
    <>
